requirements.txt
streamlit==1.28.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3